requirements.txt
# Telegram Bot (versión específica para evitar conflictos)
python-telegram-bot[job-queue]==20.3

# APIs y HTTP (compatible con telegram-bot)
httpx==0.24.1  # Versión específica compatible
aiohttp==3.9.0
requests==2.31.0

# Audio processing (OPCIONAL - comentado por ahora)
# pydub==0.25.1
# SpeechRecognition==3.10.0

# Data processing (OPCIONAL)
# pandas==2.0.3
# openpyxl==3.1.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Utilities
aiofiles==23.2.1
pytz==2023.3
tzlocal==5.2
loguru==0.7.2

# Whisper CPU only (OPCIONAL - requiere muchas dependencias)
# openai-whisper==20231117

# Para Groq API (OPCIONAL)
# groq==0.9.0
